Wrap app content in a container and add service to control
