+++ EtherGizmos.SqlMonitor.Web/src/app/app.routes.ts
@@ -1,5 +1,7 @@
 import { Routes } from "@angular/router";
 
 export const APP_ROUTES: Routes = [
-  { path: '', pathMatch: 'full', loadChildren: () => import('./features/dashboard/dashboard.routes').then(m => m.DASHBOARD_ROUTES) }
+  { path: '', pathMatch: 'full', redirectTo: 'dashboards' },
+  { path: 'dashboard', loadChildren: () => import('./features/dashboard/dashboard.routes').then(m => m.DASHBOARD_ROUTES) },
+  { path: 'dashboards', loadChildren: () => import('./features/dashboard/dashboard.routes').then(m => m.DASHBOARD_ROUTES) },
 ];
