+++ EtherGizmos.SqlMonitor.Web/src/app/features/dashboard/components/dashboard-list/dashboard-list.component.html
@@ -1,0 +1,5 @@
+<div class="row">
+  <div class="col-12">
+
+  </div>
+</div>
