--- EtherGizmos.SqlMonitor.Web/src/app/features/dashboard/components/index/dashboard.component.html
@@ -1,19 +1,0 @@
-<gridstack [options]="gridOptions" (addedCB)="onAdd($event)" (changeCB)="onChange($event)" (removedCB)="onRemove($event)">
-  <ng-container *ngIf="dashboardForm">
-    <gridstack-item *ngFor="let item of getAllFormValues(dashboardForm).items; trackBy: identifyWidget" [options]="getGridstackWidget(item)" (mouseenter)="item.grid.hovering = true" (mouseleave)="item.grid.hovering = false">
-      <ng-container [ngSwitch]="item.type">
-        <chart-widget *ngSwitchCase="'chart'"
-                      [config]="item"
-                      [startTime]="startTime"
-                      [endTime]="endTime"
-                      [instances]="instanceFilters"
-                      (onUpdate)="updateWidget($event)"
-                      (onDelete)="deleteWidget($event)"></chart-widget>
-        <text-widget *ngSwitchCase="'text'"
-                     [config]="item"
-                     (onUpdate)="updateWidget($event)"
-                     (onDelete)="deleteWidget($event)"></text-widget>
-      </ng-container>
-    </gridstack-item>
-  </ng-container>
-</gridstack>
