+++ EtherGizmos.SqlMonitor.Web/src/app/shared/services/body/body.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { BodyService } from './body.service';
+
+describe('BodyService', () => {
+  let service: BodyService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(BodyService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
